Move remainder and remquo to own testing file
