+++ test/Jamfile
@@ -103,6 +103,7 @@ run test_log.cpp ;
 run test_log1p.cpp ;
 run test_pow.cpp ;
 run test_promotion.cpp ;
+run test_remainder_remquo.cpp ;
 run test_sin_cos.cpp ;
 run test_sinh.cpp ;
 run test_sqrt.cpp ;
