+++ test/test_cmath.cpp
@@ -441,86 +441,6 @@ void test_modf()
 }
 
 template <typename Dec>
-void test_remainder()
-{
-    std::uniform_real_distribution<float> dist(-1e3F, 1e3F);
-
-    constexpr auto max_iter {std::is_same<Dec, decimal128>::value ? N / 4 : N};
-    for (std::size_t n {}; n < max_iter; ++n)
-    {
-        const auto val1 {dist(rng)};
-        const auto val2 {dist(rng)};
-        Dec d1 {val1};
-        Dec d2 {val2};
-
-        auto ret_val {std::remainder(val1, val2)};
-        auto ret_dec {static_cast<float>(remainder(d1, d2))};
-
-        if (!BOOST_TEST(std::fabs(ret_val - ret_dec) < 0.005))
-        {
-            // LCOV_EXCL_START
-            std::cerr << "Val 1: " << val1
-                      << "\nDec 1: " << d1
-                      << "\nVal 2: " << val2
-                      << "\nDec 2: " << d2
-                      << "\nRet val: " << ret_val
-                      << "\nRet dec: " << ret_dec
-                      << "\nFloat dist: " << boost::math::float_distance(ret_val, ret_dec) << std::endl;
-            // LCOV_EXCL_STOP
-        }
-    }
-
-    BOOST_TEST(isnan(remainder(std::numeric_limits<Dec>::infinity() * Dec(dist(rng)), Dec(1))));
-    BOOST_TEST(isnan(remainder(std::numeric_limits<Dec>::quiet_NaN() * Dec(dist(rng)), Dec(1))));
-    BOOST_TEST(isnan(remainder(Dec(1), std::numeric_limits<Dec>::quiet_NaN() * Dec(dist(rng)))));
-    BOOST_TEST(isnan(remainder(Dec(1), Dec(0))));
-}
-
-template <typename Dec>
-void test_remquo()
-{
-    std::uniform_real_distribution<float> dist(-1e3F, 1e3F);
-
-    constexpr auto max_iter {std::is_same<Dec, decimal128>::value ? N / 4 : N};
-    for (std::size_t n {}; n < max_iter; ++n)
-    {
-        const auto val1 {dist(rng)};
-        const auto val2 {dist(rng)};
-        Dec d1 {val1};
-        Dec d2 {val2};
-        int flt_int {};
-        int dec_int {};
-
-        auto ret_val {std::remquo(val1, val2, &flt_int)};
-        auto ret_dec {static_cast<float>(remquo(d1, d2, &dec_int))};
-
-        if ((!BOOST_TEST(std::fabs(ret_val - ret_dec) < 0.005) &&
-             (BOOST_TEST(flt_int == dec_int))))
-        {
-            // LCOV_EXCL_START
-            std::cerr << "Val 1: " << val1
-                      << "\nDec 1: " << d1
-                      << "\nVal 2: " << val2
-                      << "\nDec 2: " << d2
-                      << "\nRet val: " << ret_val
-                      << "\nRet dec: " << ret_dec
-                      << "\nInt val: " << flt_int
-                      << "\nInt quo: " << val1 / val2
-                      << "\nInt dec: " << dec_int
-                      << "\nDec quo: " << d1 / d2
-                      << "\nFloat dist: " << boost::math::float_distance(ret_val, ret_dec) << std::endl;
-            // LCOV_EXCL_STOP
-        }
-    }
-
-    int quo {};
-    BOOST_TEST(isnan(remquo(std::numeric_limits<Dec>::infinity() * Dec(dist(rng)), Dec(1), &quo)));
-    BOOST_TEST(isnan(remquo(std::numeric_limits<Dec>::quiet_NaN() * Dec(dist(rng)), Dec(1), &quo)));
-    BOOST_TEST(isnan(remquo(Dec(1), std::numeric_limits<Dec>::quiet_NaN() * Dec(dist(rng)), &quo)));
-    BOOST_TEST(isnan(remquo(Dec(1), Dec(0), &quo)));
-}
-
-template <typename Dec>
 void test_fdim()
 {
     std::uniform_real_distribution<float> dist(1.0F, 1e5F);
@@ -1561,13 +1481,6 @@ int main()
 
     test_modf<decimal32>();
     test_modf<decimal64>();
-
-    test_remainder<decimal32>();
-    test_remquo<decimal32>();
-    test_remainder<decimal64>();
-    test_remquo<decimal64>();
-    test_remainder<decimal128>();
-    test_remquo<decimal128>();
 
     test_fdim<decimal32>();
     test_fdim<decimal64>();
